Finish Repository

+++ src/main/java/com/el/ally/invoicify/models/Company.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
@@ -16,18 +17,18 @@ public class Company {
   @Id
   @GeneratedValue(generator="people_id_seq", strategy=GenerationType.AUTO)
   
-  private Long id;
+  private int id;
 
   private String name;
 
   @OneToMany(mappedBy="company")
   private List<Invoice> invoices;
 
-public Long getId() {
+public int getId() {
 	return id;
 }
 
-public void setId(Long id) {
+public void setId(int id) {
 	this.id = id;
 }
